rename filenames mentioning parties
