+++ src/snarky/gen/dune
@@ -1,5 +1,5 @@
 (rule
- (targets js-layout.ts parties.ts parties-json.ts)
+ (targets js-layout.ts transaction.ts transaction-json.ts)
  (mode promote)
  (deps
   (:script ../../build/jsLayoutToTypes.mjs)
